--- apps/meteor/ee/client/apps/index.ts
@@ -1,3 +1,0 @@
-import './gameCenter/tabBar';
-
-export { AppClientOrchestratorInstance as Apps } from './orchestrator';
