+++ apps/meteor/ee/client/startup/index.ts
@@ -1,4 +1,3 @@
-import '../apps';
 import './audit';
 import './deviceManagement';
 import './slashCommands';
